==> scripts/04_region_inclusive.R <==
library(here)
library(tidyverse)

regional <- read_csv("./processeddata/species_lists/region/20230410_fish_by_region.csv")
top500 <- read_csv("./processeddata/species_lists/region/20230410_fish_top500_taxized.csv")
