==> scripts/04a_ncbi-search.Rmd <==
---
title: "04_mifish-blast"
output: html_document
date: "2023-05-02"
---

This script will prepare ... stufff ... for ncbi



```{r setup, include=FALSE}

library(tidyverse)
library(here)

splists <- here('processeddata', 'species_lists')
```

```{r load-in}
# fish
region_fish <- read_csv(here(splists, 'region', '20230501_obis-fish_fishbase_nep.csv'))[-1] 

# list of fish in mitofish
mitofish_fish <- read_csv(here(splists, 'ncbi', '20230511_mitofish-species.csv')) %>%
  distinct(uid, .keep_all = TRUE)

#region_inverts <- read_csv()
```


# Make search

## Fish

We need to generate a few outputs for future steps.

- A species list in the format "Genus_species" each on a new line
- A list of taxid's in a format that can be taken by an NCBI Entrez search or BLAST.

```{r generate species list}
fishlist <- region_fish %>%
  select(accepted_name) %>%
  distinct() %>%
  # turn into genus_species format
  mutate(accepted_name = str_replace_all(accepted_name, " ", "_"))

# write delim will just write the txt file, 
# col_names = FALSE to remove anything that isn't a species name
write_delim(fishlist, here(splists, 'region', '20230511_fishsp.txt'), 
            col_names = FALSE)
```

This contains all species in our region
```{r full ncbi-query}
# determine which fish should be searched
# 
region_fish_search2 <- region_fish %>%
  filter(!(is.na(ncbi_id))) %>%
  #filter() by taxonomy, region, etc if needed
  # then select levels filtered
  select(ncbi_id, accepted_name) %>%
  distinct() %>%
  # what is chr?
  mutate(chr = paste("txid", ncbi_id, "[Orgn]", sep = ""))

region_fish_search_str <- toString(region_fish_search2$chr) %>%
  gsub(' ', '',.)

# paste into search here: https://www.ncbi.nlm.nih.gov/nuccore/advanced
# or save
write_file(region_fish_search_str, file = here(splists, 'ncbi', '20230501_region-fish-ncbisearch.txt'))

```

This filters the species in our region by presence in mitofish before writing
Mitofish list generated from extracting accessions and UIDs from the mitofish database pulled from CRABS.
```{r filtered mitofish ncbi}

# determine which fish should be searched
# 
region_nomitofish <- region_fish %>%
  filter(!(is.na(ncbi_id))) %>%
  #filter() by taxonomy, region, etc if needed
  filter(!(ncbi_id %in% mitofish_fish$uid)) %>%
  # then select levels filtered
  select(ncbi_id, accepted_name) %>%
  distinct() %>%
  # what is chr?
  mutate(chr = paste("txid", ncbi_id, "[Orgn]", sep = ""))

region_fish_search_str <- toString(region_nomitofish$chr) %>%
  gsub(' ', '',.)

# paste into search here: https://www.ncbi.nlm.nih.gov/nuccore/advanced
# or save
write_file(region_fish_search_str, file = here(splists, 'ncbi', '20230501_region-fish-ncbisearch-nomitofish.txt'))

```





# scratch

ben's script
```{r}
#messing around
#test datasets on few sspecies
prop <- readDNAStringSet("./rawdata/NCBI_searches/PROP.fasta")
word <- readDNAStringSet("./rawdata/NCBI_searches/WORD.fasta")
mito <- readDNAStringSet("./rawdata/NCBI_searches/mito.fasta")

seq_name_p = names(prop)
seq_name_w = names(word)
seq_name_m = names(mito)

accessions_p <- sapply(strsplit(seq_name_p, " "), "[", 1)
accessions_w <- sapply(strsplit(seq_name_w, " "), "[", 1)
accessions_m <- sapply(strsplit(seq_name_m, " "), "[", 1)

sequence = paste(prop)
dataframe <- data.frame(accessions_p, seq_name_p, sequence)

t1 <- accessionToTaxa(dataframe$accessions_p)
accessionToTaxa(c("Z17430.1","Z17429.1","X62402.1",'NOTREAL'),sqlFile)






#sequence_p = paste(prop)
#dataframe_p <- data.frame(seq_name, sequence)

accessions_p <- sapply(strsplit(seq_name_p, " "), "[", 1)
accessions_w <- sapply(strsplit(seq_name_w, " "), "[", 1)
accessions_m <- sapply(strsplit(seq_name_m, " "), "[", 1)

pw <- accessions_p[!(accessions_p %in% accessions_w)]
pm <- accessions_p[!(accessions_p %in% accessions_m)]
wm <- accessions_w[!(accessions_w %in% accessions_m)]
mp <- accessions_m[!(accessions_m %in% accessions_p)]









#inverts
inverts <- read.csv('./processeddata/species_lists/region/20230412_obis-invertebrates_nep.csv')%>%
  select(c("ncbi_id")) %>%
  distinct()

b1 <- inverts %>%
  mutate(chr = paste("txid", ncbi_id, "[ORGN]", sep = ""))
b2 <- toString(b1$chr)

b3 <- gsub(',', ' or',b2)
b4 <- paste("(",b3,"AND mitochondrial [WORD] NOT UNVERIFIED [WORD] NOT PREDICTED [WORD] AND 100:20000 [SLEN]")
b4 


```
